% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_precinct_map.R
\name{create_precinct_map}
\alias{create_precinct_map}
\title{Create Precinct Map}
\usage{
create_precinct_map(d)
}
\arguments{
\item{d}{A geojson file}
}
\value{

}
\description{
Create Precinct Map
}
