% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_precinct_summary.R
\name{create_precinct_summary}
\alias{create_precinct_summary}
\title{Create Precinct Summary}
\usage{
create_precinct_summary(d, precinct_filter)
}
\arguments{
\item{d}{A data set containing police precincts}

\item{precinct_filter}{Precinct to filter data}
}
\value{
return summarized data set
}
\description{
Create Precinct Summary
}
