% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_bi_bar_graph.R
\name{make_bi_bar_graph}
\alias{make_bi_bar_graph}
\title{Title}
\usage{
make_bi_bar_graph(
  d,
  var1,
  var2,
  label1,
  label2,
  labels = F,
  voption = "plasma"
)
}
\arguments{
\item{voption}{}
}
\value{

}
\description{
Title
}
